+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-# V1.2.0 - 12 May 2022
+# V1.2.1 - 13 Jul 2023
+	- Ordena por folio la lista de corte
+		- https://app.asana.com/0/1202827617117921/1204952471293765/f
+	- Incluye la fecha de venta en el modulo de corte y modificar ventas
+		- https://app.asana.com/0/1202827617117921/1204952471293767/f
+
+# V1.2.0 - 12 May 2023
 	- Upgrade webpack to `5.54.0`
 	- Delete engines from package.json
 	- Add hashFunction to webpack.config
@@ -12,7 +18,7 @@
 # V1.0.0 - 05 Oct 2022
 	- New data base for testing added
 		- https://app.asana.com/0/1202827617117921/1202875997801329/f
-	-  Reasignation error for seat 2 fixed
+	- Reasignation error for seat 2 fixed
 		- https://app.asana.com/0/1202827617117921/1202875997801296/f
 	- New feature: Viajes user can view the corte for all employees, but cannot modify its status
 		- https://app.asana.com/0/1202827617117921/1202875997801286/f